+++ js-miniapp-sdk/test/miniapp.spec.ts
@@ -80,6 +80,9 @@ window.MiniAppBridge = {
     getPermissionStatus: sandbox.stub(),
     launchAppSettings: sandbox.stub(),
   },
+  browserManager: {
+    loadUsingHTMLString: sandbox.stub(),
+  },
 };
 
 const miniApp = new MiniApp();
@@ -987,5 +990,35 @@ describe('launchAppSettings', () => {
     );
     return expect(miniApp.miniappUtils.launchAppSettings()).to.eventually.be
       .rejected;
+  });
+});
+
+describe('loadUsingHTMLString', () => {
+  const params = {
+    htmlString: '<html><body>Hello World</body></html>',
+    callbackUrl: 'https://localhost:3000/index.html',
+    baseUrl: 'https://localhost',
+  };
+
+  it('should return if load using HTML String', () => {
+    window.MiniAppBridge.browserManager.loadUsingHTMLString.resolves(true);
+    return expect(
+      miniApp.miniappUtils.loadUsingHTMLString(
+        params.htmlString,
+        params.callbackUrl
+      )
+    ).to.eventually.equal(true);
+  });
+
+  it('should return error information', () => {
+    window.MiniAppBridge.browserManager.loadUsingHTMLString.returns(
+      Promise.reject('test error')
+    );
+    return expect(
+      miniApp.miniappUtils.loadUsingHTMLString(
+        params.htmlString,
+        params.callbackUrl
+      )
+    ).to.eventually.be.rejected;
   });
 });